Turbo is black
